chatops/tests/e2e/contract/provider/test_engine_provider.py
"""
Pact Provider Verification - Engine as Provider

These tests verify that Engine API fulfills the contracts defined by consumers.
"""

import pytest
import subprocess
from pact import Verifier


class TestEngineProviderVerification:
    """Verify Engine API against consumer contracts."""

    @pytest.fixture
    def provider_url(self):
        """Engine provider URL for verification."""
        return "http://localhost:8000"

    @pytest.fixture
    def pact_broker_url(self):
        """Pact Broker URL."""
        return "http://localhost:9292"

    def test_verify_gateway_contract(self, provider_url):
        """Verify contracts defined by Gateway consumer."""
        verifier = Verifier(
            provider="ChatOpsEngine",
            provider_base_url=provider_url,
        )

        # Verify against local pact file
        output, logs = verifier.verify_pacts(
            "./pacts/chatopsgateway-chatopsengine.json",
            provider_states_setup_url=f"{provider_url}/_pact/provider-states",
        )

        assert output == 0, f"Pact verification failed: {logs}"

    def test_verify_from_broker(self, provider_url, pact_broker_url):
        """Verify contracts from Pact Broker."""
        verifier = Verifier(
            provider="ChatOpsEngine",
            provider_base_url=provider_url,
        )

        try:
            output, logs = verifier.verify_with_broker(
                broker_url=pact_broker_url,
                enable_pending=True,
                include_wip_pacts_since="2024-01-01",
                provider_states_setup_url=f"{provider_url}/_pact/provider-states",
                publish_verification_results=True,
                provider_app_version="1.0.0",
            )

            assert output == 0, f"Pact verification failed: {logs}"
        except Exception as e:
            pytest.skip(f"Pact Broker not available: {e}")


class TestEngineProviderStates:
    """Provider state setup for Pact verification."""

    PROVIDER_STATES = {
        "Engine is available": {
            "setup": lambda: None,  # No setup needed
            "teardown": lambda: None,
        },
        "Engine is available for batch processing": {
            "setup": lambda: None,
            "teardown": lambda: None,
        },
        "A processing job exists": {
            "setup": lambda: _create_test_job("exec-123"),
            "teardown": lambda: _cleanup_test_job("exec-123"),
        },
        "Engine is running": {
            "setup": lambda: None,
            "teardown": lambda: None,
        },
        "Engine validates input": {
            "setup": lambda: None,
            "teardown": lambda: None,
        },
    }

    @classmethod
    def setup_state(cls, state_name: str):
        """Set up the given provider state."""
        if state_name in cls.PROVIDER_STATES:
            cls.PROVIDER_STATES[state_name]["setup"]()
        else:
            raise ValueError(f"Unknown provider state: {state_name}")

    @classmethod
    def teardown_state(cls, state_name: str):
        """Tear down the given provider state."""
        if state_name in cls.PROVIDER_STATES:
            cls.PROVIDER_STATES[state_name]["teardown"]()


def _create_test_job(job_id: str):
    """Create a test processing job."""
    # In real implementation, this would create test data
    pass


def _cleanup_test_job(job_id: str):
    """Clean up test processing job."""
    # In real implementation, this would clean up test data
    pass


# Provider State Handler for use with web framework
"""
Example Flask implementation for provider states:

from flask import Flask, request, jsonify

app = Flask(__name__)

@app.route("/_pact/provider-states", methods=["POST"])
def provider_states():
    state = request.json.get("state")
    action = request.json.get("action", "setup")

    handler = TestEngineProviderStates.PROVIDER_STATES.get(state)
    if handler:
        if action == "setup":
            handler["setup"]()
        else:
            handler["teardown"]()
        return jsonify({"status": "ok"})

    return jsonify({"error": f"Unknown state: {state}"}), 400
"""
