chatops/tests/e2e/requirements.txt
# E2E Testing Dependencies

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
pytest-timeout>=2.1.0
pytest-html>=4.0.0

# HTTP Client
httpx>=0.24.0
requests>=2.31.0
aiohttp>=3.8.0

# Contract Testing
pact-python>=2.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0

# Redis
redis>=4.6.0
aioredis>=2.0.0

# Data Validation
pydantic>=2.0.0
faker>=18.0.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
structlog>=23.1.0

# Async Support
anyio>=3.7.0

# JSON Schema
jsonschema>=4.18.0
